Pluggable ambiance
Made the ambient unit of work storage pluggable, so it does no longer
default to ThreadStatic. In the course of doing that, I undermined the
simplicity.  On the upside the middle man between the repository and the
unit of work (scope) has been made explicit. The future of this piece of
code is uncertain. DI containers supporting scoping/child containers
make this scenario way easier, yet I'd rather have something in the box
and a plugin for every DI container is not it. Am I looking for an
abstract factory? I think I'll need to consume this code more to be
become aware of the path forward. Design woes, hate it, love it.