As a matter of facts
- Adding "facts" to the testing API so specifications can be composed
using regular methods instead of the extension methods or reverting to
inheritance hackery. It's the composition that aids in easing the
readability. After all, you can tuck away a whole story behind a
descriptive method name in the context of a specification. This
resembles what happens in regular TDD where you want to minimize the
setup.